roastmymvp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "roastmymvp"
version = "0.2.0"
description = "AI-simulated beta testing + VC roast for any product"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "click>=8.1",
    "playwright>=1.40",
    "anthropic>=0.40",
    "pyyaml>=6.0",
]

[project.scripts]
roastmymvp = "roastmymvp.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["roastmymvp"]

[tool.coverage.report]
fail_under = 80
